# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "409": {
            "description": "Email déjà utilisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un nouvel utilisateur",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer tous les utilisateurs",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un utilisateur par son ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un utilisateur",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur supprimé"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un utilisateur",
        "tags": [
          "users"
        ]
      }
    },
...
    "/deadlines/user/{userId}": {
      "get": {
        "operationId": "DeadlinesController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des échéances par utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deadline"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les échéances par utilisateur",
        "tags": [
          "deadlines"
        ]
      }
    },
...
    "/projects/team/{teamId}": {
      "get": {
        "operationId": "ProjectsController_findByTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "description": "ID de l'équipe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des projets de l'équipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les projets par équipe",
        "tags": [
          "projects"
        ]
      }
    },
    "/teams": {
      "post": {
        "operationId": "TeamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Équipe créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une nouvelle équipe",
        "tags": [
          "teams"
        ]
      },
      "get": {
        "operationId": "TeamsController_findAll",
        "parameters": [
          {
            "name": "leaderId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des équipes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer toutes les équipes",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{id}": {
      "get": {
        "operationId": "TeamsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'équipe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Équipe trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "404": {
            "description": "Équipe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer une équipe par son ID",
        "tags": [
          "teams"
        ]
      },
      "patch": {
        "operationId": "TeamsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'équipe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Équipe mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "404": {
            "description": "Équipe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une équipe",
        "tags": [
          "teams"
        ]
      },
      "delete": {
        "operationId": "TeamsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'équipe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Équipe supprimée"
          },
          "404": {
            "description": "Équipe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une équipe",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/members/{userId}": {
      "post": {
        "operationId": "TeamsController_addMember",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "description": "ID de l'équipe",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur à ajouter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membre ajouté avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "404": {
            "description": "Équipe ou utilisateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajouter un membre à une équipe",
        "tags": [
          "teams"
        ]
      },
      "delete": {
        "operationId": "TeamsController_removeMember",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "description": "ID de l'équipe",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur à retirer",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membre retiré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "404": {
            "description": "Équipe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retirer un membre d'une équipe",
        "tags": [
          "teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestion d'Échéances",
    "description": "API pour l'application de gestion d'échéances du secteur défense",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Prénom",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "description": "Nom",
            "example": "Dupont"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "jean.dupont@example.com"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe",
            "example": "P@ssw0rd123"
          },
          "role": {
            "type": "string",
            "description": "Rôle de l'utilisateur",
            "enum": [
              "admin",
              "manager",
              "user"
            ],
            "default": "user"
          },
          "department": {
            "type": "string",
            "description": "Département",
            "example": "Ingénierie"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "User": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Prénom",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "description": "Nom",
            "example": "Dupont"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "jean.dupont@example.com"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe",
            "example": "P@ssw0rd123"
          },
          "role": {
            "type": "string",
            "description": "Rôle de l'utilisateur",
            "enum": [
              "admin",
              "manager",
              "user"
            ],
            "default": "user"
          },
          "department": {
            "type": "string",
            "description": "Département",
            "example": "Ingénierie"
          }
        }
      },
...
      "Deadline": {
        "type": "object",
        "properties": {}
      },
...
      "Project": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom de l'équipe",
            "example": "Équipe Cybersécurité"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée de l'équipe",
            "example": "Équipe responsable de la sécurité des systèmes d'information"
          },
          "leaderId": {
            "type": "string",
            "description": "ID de l'utilisateur chef d'équipe",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "memberIds": {
            "description": "Liste des IDs des membres de l'équipe",
            "example": [
              "7a9d0f35-d562-4219-c732-d35cf1a2e5f9",
              "3e7d0f55-e987-4215-c382-a56cf1a2e4d8"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "department": {
            "type": "string",
            "description": "Département auquel l'équipe appartient",
            "example": "R&D"
          }
        },
        "required": [
          "name",
          "leaderId",
          "memberIds"
        ]
      },
      "Team": {
        "type": "object",
        "properties": {}
      },
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom de l'équipe",
            "example": "Équipe Cybersécurité"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée de l'équipe",
            "example": "Équipe responsable de la sécurité des systèmes d'information"
          },
          "leaderId": {
            "type": "string",
            "description": "ID de l'utilisateur chef d'équipe",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "memberIds": {
            "description": "Liste des IDs des membres de l'équipe",
            "example": [
              "7a9d0f35-d562-4219-c732-d35cf1a2e5f9",
              "3e7d0f55-e987-4215-c382-a56cf1a2e4d8"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "department": {
            "type": "string",
            "description": "Département auquel l'équipe appartient",
            "example": "R&D"
          }
        }
      },
...
    }
  }
}